Move init method to generators folder
