+++ kybra/compiler/kybra_generate/src/generators/canister_methods/mod.rs
@@ -1,1 +1,2 @@
 pub mod heartbeat;
+pub mod init;
